backend/pyproject.toml
[tool.poetry]
name = "feedback"
version = "0.1.0"
description = "feedback backend module"
authors = ["Vernadskii <d.vernadskii@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
django-ninja = "^1.2.2"

[tool.poetry.group.dev.dependencies]
coverage = "^7.2.7"
flake8-annotations = "^3.0.1"
flake8-datetime-import = "^0.1.0"
flake8-new-union-types = "^0.4.1"
flake8-pep585 = "^0.1.7"
flake8-print = "^5.0.0"
isort = "^5.12.0"
wemake-python-styleguide = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
